LPS-153813 LPS-166948 search-experiences-web: Add test for isElementInactiveFromNonCompanyIndex
