+++ modules/dxp/apps/search-experiences/search-experiences-web/test/js/edit_sxp_blueprint/add_sxp_element_sidebar/index.js
@@ -13,7 +13,7 @@ import {fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
 import AddSXPElementSidebar from '../../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/add_sxp_element_sidebar/index';
-import {QUERY_SXP_ELEMENTS} from '../../mocks/data';
+import {QUERY_SXP_ELEMENTS} from '../../mocks/sxpElements';
 
 import '@testing-library/jest-dom/extend-expect';
