+++ modules/dxp/apps/search-experiences/search-experiences-web/test/js/edit_sxp_blueprint/query_builder_tab/index.js
@@ -13,7 +13,7 @@ import {fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
 import QueryBuilder from '../../../../src/main/resources/META-INF/resources/sxp_blueprint_admin/js/edit_sxp_blueprint/query_builder_tab/index';
-import {QUERY_SXP_ELEMENTS} from '../../mocks/data';
+import {QUERY_SXP_ELEMENTS} from '../../mocks/sxpElements';
 
 import '@testing-library/jest-dom/extend-expect';
